pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "projectg"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "selenium",
    "python-dotenv",
    "prisma"
]
authors = []

[tool.setuptools]
package-dir = { "" = "backend" }

[tool.setuptools.packages.find]
where = ["backend"]
include = ["project_g*"]
exclude = ["project_g.dependencies*"]